http serializer work

+++ library/http/include/library/http/version.h
@@ -5,8 +5,8 @@
 namespace NHttp {
 
     struct THttpVersion {
-        uint8_t major = 1;
-        uint8_t minor = 1;
+        uint16_t major = 1;
+        uint16_t minor = 1;
         
         bool operator==(const THttpVersion& other) const noexcept;
         bool operator!=(const THttpVersion& other) const noexcept;
